feat: transition to ts
