+++ .gitignore
@@ -14,3 +14,6 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+dist
+package-lock.json
