--- index.js
@@ -1,9 +1,0 @@
-import Jimp from 'jimp';
-import {httpServer} from './src/http_server/index.js';
-import robot from 'robotjs';
-import { WebSocketServer } from 'ws';
-
-const HTTP_PORT = 3000;
-
-console.log(`Start static http server on the ${HTTP_PORT} port!`);
-httpServer.listen(HTTP_PORT);
