+++ package.json
@@ -2,20 +2,32 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start:prod": "npm run build && npm run prod",
+        "start:dev": "nodemon src/index.ts",
+        "prod": "node dist/index.js",
+        "build": "tsc",
+        "test": "jest"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/jest": "^28.1.2",
+        "@types/node": "^17.0.42",
+        "@types/supertest": "^2.0.12",
+        "@types/uuid": "^8.3.4",
+        "@types/ws": "^8.5.3",
+        "cross-env": "^7.0.3",
+        "jest": "^28.1.1",
+        "nodemon": "^2.0.16",
+        "supertest": "^6.2.3",
+        "ts-jest": "^28.0.5",
+        "ts-node": "^10.8.1",
+        "typescript": "^4.7.3"
     },
     "dependencies": {
+        "cross-env": "^7.0.3",
         "jimp": "^0.16.1",
         "robotjs": "^0.6.0",
         "ws": "^8.8.0"
